test(pvz): add getPvz tests
